Add system Groovy scripts to dynamically manage nodes
When we run jobs using the current CI infrastrucutre, we create new
nodes and run tasks on them, but we do not register them as Jenkins
nodes, so we cannot use any of the first-class node-addressable act-
ions that Jenkins is capable of. By dynamically adding and removing
nodes as our jobs run, our job DSL can be much more idiomatic.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>