Use one generic pod template for tests
Exposing a simple interface to users adding tests requires that we DRY
out the boilerplate that we use to run them. The first step is to use
one template to create the Pods we run the tests in. These can simply be
parameterized with the name of the test, command-line to run and the
quota allocation.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>