prow: Remove old branched openshift/os:latest build
We're moving to a container in the rhcos/ namespace; see
https://github.com/openshift/release/issues/972

This is part of an effort to consolidate the "source of truth" for RHCOS content.
We'd like to get to one container, and short term, having it built from
the same process that builds other content internally, and pushed to the osorgci registry
makes things far easier.