Add a ci-operator Prow job generator
The purpose of this tool is to reduce an amount of boilerplate that
component owners need to write when they use ci-operator to set up CI
for their component. The generator is able to entirely generate the
necessary Prow job configuration from the ci-operator configuration
file.

The generator creates one presubmit job for each test specified in the
ci-operator config file (in tests list). Also, if the configuration file
has a non-empty images list, one additional presubmit and postsubmit job
is generated with --target=[images] option passed to ci-operator to
attempt to build the component images. This postsubmit job also uses the
--promote option to promote the component images built in this way.

The generator can use the naming convention and directory structure of
this repository. Using this mode, all the user needs to do is to place
the ci-operator configuration file to the correct place in
ci-operator/config directory and then run the generator with --full-repo
parameter. With --full-repo option, the generator uses git to detect a
root directory of the repo and then it constructs paths to both
directories from it. It is also possible to not use the auto-detection
and pass the paths directly.