cluster/ci/config/prow/plugins: Enable approve and blunderbuss
We grew OWNERS files for a number of directories in 8ebd3290
(Auto-generate OWNERS for config, 2018-08-16, #1201).  And bfabbce
(OWNERS: Add approvers for the project as a whole, 2018-08-20) gave us
a backstop for any content that does not have a more-specific OWNERS
file.  With this commit, we start using those OWNERS files to gate
approval.