Machine-write all Prow job config YAMLs
This should have no impact on *any* job definition. It just rewrites all
YAML in syntactically different, but semantically equivalent way.

We are building tooling to generate Prow job configuration and merge it
with existing jobs, which will result in marshaling job configs into
existing files. To prepare for using these tools, this commit rewrites
all configs with the same YAML marshal lib, so that changes done by the
tools are better isolated and readable.