openshift-installer-presubmits: Add go-fmt
We've had some gofmt issues sneak in recently [1].  We want to catch
that sort of thing in CI, instead of finding out about it later.  This
commit uses the script from openshift/installer@94a2a8ca (hack/go-fmt:
Make it easy to auto-format Go, 2018-08-24, openshift/installer#173)
to automate that check.

[1]: https://github.com/openshift/installer/commit/8f415b93d638966481d616d2cdb983cc3fde9b4f#commitcomment-30284820