cluster/test-deploy/aws/openshift: Set the pull secret as a string
Catching up with openshift/installer@8a84eb2e (*: Replace
PullSecretPath with PullSecret, 2018-09-05, openshift/installer#208).

The jq call ensures the secret JSON is on a single line, otherwise
we'd need to be a bit more careful about injecting it into the YAML as
a valid string literal.

I've replaced the inputs.yaml copy with a sed invocation to avoid
leaking the internal pull secret into the output artifacts.  I'm not
sure how sensitive it is though, maybe we don't mind leaking the
secret?