openshift/os: use coreos-assembler container and cleanup
We're standardizing on the coreos-assembler container for all our CI/CD
needs. Let's also make sure our PR tests are updated to match.

I deleted the second presubmit job since it's not doing anything
right now (the target Dockerfile is just empty). We'll likely revive it
at some point to also use coreos-assembler (or just merge it as part of
the first presubmit job), but let's do that in a separate PR.