cluster-launch-installer-e2e*: Drop AMI override
Since openshift/installer@66ef3beb (*: remove support for container
linux, 2018-09-06, openshift/installer#221), the installer defaults to
a recent RHCOS AMI.  Drop the override here, so we can track RHCOS
with only installer-repo bumps.  I've left the variable in the
template so that, if this repo needs to have its own, separate,
opinion again, it can set the override again.  However, to support the
no-local-opinion case, the new local-opinion syntax will be:

  export EC2_AMI_OVERRIDE='ec2AMIOverride: ami-07307c397daf4d02e'