ci-operator/config/openshift/installer/master: Move to openshift-install
openshift-install is he next-gen installer.  With the new installer,
we only need the installer binary and terraform to launch and destroy
clusters.  I've removed the smoke tests for now, because I haven't
looked into building them without Bazel.  Hopefully we'll get expanded
e2e testing *and* other OpenShift projects using our installer soon to
keep us honest.

Also missing from the new installer is a way to set expirationDate.
But we'll just not leak until we regain the ability to set that,
right? ;)

The new installer dumps less cruft into the output directory (most of
the generated output goes into a temporary directory), so I've
adjusted the openshift-install calls to just use an artifacts subdir
for their state storage.  A bonus of this approach is that if the
installer hangs up, we'll capture anything it wrote to disk without
needing an explicit cp call.  A drawback is that we'll leak any
secrets that get put into the kubeconfig, so don't put anything in
there that needs to stay private after the cluster is reaped.

Using the base image for the installer's 'from' is quite a bit
different from openshift/installer@29e4d10e (origin/pr/343)
images/installer: Rewrite tectonic-installer for openshift-install,
2018-09-26, openshift/installer#343), where the Dockerfile is 'FROM
scratch'.  Including the OpenShift base currently adds ~230 MB to the
installer layer's 110 MB for the two binaries, although both of those
are uncompressed sizes.  Gzipping layers reduces the sizes to around
84 MB and 26 MB respectively.  So the added base cruft is not huge,
but it's still hefty.

The upside of using the base image is that we have a standard
POSIX-ish system for executing the cleanup script.  To demonstrate the
'FROM scratch' approach, I've adjusted the setup container to call the
installer directly (with no wrapping shell script).  And to support
that, I've shifted some waiting code over into the test container
(which also saves us from having to copy 'oc' around).

I've also dropped the 3.11 config, since the installer is 4.0-only.