add simple test suite for Heketi to openshift-ci
This patch adds the first (hopefully of many) test of a gluster family
component within the openshift CI system. Right now I'm trying to
start very simple with only an equivalent of what we're doing in
travis runs in the Heketi repo. Eventually, we plan on expanding on that
within Heketi's repo as well as out to other components in the
gluster-on-openshift ecosystem.

Non-goals for this PR:
* Having centos-ci/bots manage the PR workflow in Heketi
* Integration tests
* Persistent container builds

Hopefully, this is fairly close to what we need to start but for
review purposes I'd like to point out I was fuzzy if we needed
anything beyond trigger in plugins.yaml. I was also fuzzy on
the relationships between the base_images & build_root sections
in the config yaml file.

Signed-off-by: John Mulligan <jmulligan@redhat.com>