populate-owners: Attempt pulls over HTTP
As discussed in #1601, pulling all projects serially is slow, because
even with shallow clone you're pulling all the files in the HEAD tree.
As discussed in e1f993fb (populate-owners: Also slurp OWNERS_ALIASES,
2018-08-25, #1285), GitHub does not enable git-upload-archive, so
we can't use the Git protocol to select specific files.

One way to speed things up would be to allow fetching subsets of
projects, but that risks corrupting OWNERS_ALIASES.  If the utility
only fetched some repositories, it would have an incomplete picture of
aliases.  For example, say the repository contained
ci-operator/jobs/a/b and ci-operator/jobs/c/d, both of which defined
aliases.  A full run would result in aliases for both a/b and c/d in
OWNERS_ALIASES.  A subsequent run with only a/b would update any a/b
aliases, but all other aliases, including those which had previously
been injected by c/d, would be removed.  You could work around that by
caching information about which repositories were involved in
OWNERS_ALIASES, but it would be a bit complicated.

This commit, on the other hand, takes advantage of the fact that most
of our repositories are public.  It optimistically attempts an
efficient HTTP pull, and only falls back to the shallow clone if the
efficient pull fails.  That's slightly more work for private
repositories, where the HTTP attempt is wasted effort, but it's much
more efficient for public repos.

Another potential issue is consuming your API quota of 60
unauthenticated requests per hour [1].  But if you hit that limit, we
just fall back to Git, so it's not the end of the world.

Another alterantive would be to pull in parallel (as we used to before
e1f993fb), but that still has the bandwidth cost of shifting lots of
files that we don't care about for this use case.  You could save some
bandwidth by caching blobs locally, but then you'd have the
complication of cache management.

The "got owners ..." message is because the HTTP pulls don't produce
any terminal output.  The Git pulls write status updates to stderr,
but now that it might be a while between Git pulls, it's good to let
the caller know something is happening ;).

Docs for the commit endpoint I'm hitting are in [2].

[1]: https://developer.github.com/v3/#rate-limiting
[2]: https://developer.github.com/v3/repos/commits/#get-the-sha-1-of-a-commit-reference