cluster-launch-installer-e2e: Start setting expirationDate again
We'd dropped this in 3f2f01c4
(ci-operator/config/openshift/installer/master: Move to
openshift-install, 2018-09-26, #1677).  But since
openshift/installer@98a35316 (pkg/asset/installconfig: Add
_CI_ONLY_STAY_AWAY_OPENSHIFT_INSTALL_AWS_USER_TAGS, 2018-09-28,
openshift/installer#364), the installer supports setting it again, so
we can add it back.

As the name suggests, this variable is not a stable API.  Soon the new
installer will be able to load the configuration from YAML (again),
but we can keep adjusting as the installer evolves.