cluster-launch-installer-e2e: Pull all logs in parallel
With 10 pulls going at once.

These are currently generating a lot of error messages.  From recent
openshift/installer#415 tests [1]:

  $ oc project ci-op-w11cl72x
  $ oc logs e2e-aws -c teardown --timestamps
  2018-10-04T18:17:06.557740109Z Gathering artifacts ...
  2018-10-04T18:17:24.875374828Z Error from server (Forbidden): Forbidden (user=kube-apiserver, verb=get, resource=nodes, subresource=log)
  ...
  2018-10-04T18:17:29.331684772Z Error from server (Forbidden): Forbidden (user=kube-apiserver, verb=get, resource=nodes, subresource=log)
  2018-10-04T18:17:29.351919855Z Error from server (NotFound): the server could not find the requested resource
  2018-10-04T18:17:39.592948165Z Error from server (BadRequest): previous terminated container "registry" in pod "registry-b6df966cf-fkhpl" not found
  ...
  2018-10-04T18:29:24.457841097Z Error from server (BadRequest): previous terminated container "kube-addon-operator" in pod "kube-addon-operator-775d4c8f8d-289zm" not found
  2018-10-04T18:29:24.466213055Z Waiting for node logs to finish ...
  2018-10-04T18:29:24.466289887Z Deprovisioning cluster ...
  2018-10-04T18:29:24.483065903Z level=debug msg="Deleting security groups"
  ...
  2018-10-04T18:33:29.857465158Z level=debug msg="goroutine deleteVPCs complete"

So 12 minutes to pull the logs, followed by four minutes for
destroy-cluster.

Looking at the extracted logs, lots of them are zero (which compresses
to 20 bytes):

  $ POD_LOGS="$(w3m -dump https://gcsweb-ci.svc.ci.openshift.org/gcs/origin-ci-test/pr-logs/pull/openshift_installer/415/pull-ci-openshift-installer-master-e2e-aws/456/artifacts/e2e-aws/pods/)"
  $ echo "${POD_LOGS}" | grep '^ *20$' | wc -l
  86
  $ echo "${POD_LOGS}" | grep '\[file\]' | wc -l
  172

So it's possible that the delay is due to the errors, or to a few
large logs blocking the old, serial pod/container pulls.

With this commit, I've added a new 'queue' command.  This command
checks to see how many background jobs we have using 'jobs' [2], and
idles until we get below 10.  Then it launches its particular command
in the background.  By using 'queue', we'll keep up to 10 log-fetches
running in parallel, and the final 'wait' will block for any which
still happen to be running by that point.

The previous gzip invocations used -c, which dates back to 82d333eb
(Set up artifact reporting for ci-operator jobs, 2018-05-17, #867).
But with these gzip filters running on stdin anyway, the -c was
superfluous.  I've dropped it in this commit.

Moving redirect target to a positional argument is a bit cludgy.  I'd
rather have a more familiar way of phrasing that redirect, but passing
it in as ${1} was the best I've come up with.

[1]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/415/pull-ci-openshift-installer-master-e2e-aws/456/build-log.txt
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/jobs.html