ci-operator/jobs/infra-periodics: Opt installer out of periodic-retester
Currently things like [1,2] that try to unstick us vs. some external
change we need to /hold the other approved PRs to get them out of the
merge queue while the fix goes in.  With the bot removed from our
repository, those PRs would remove themselves as they failed
naturally, and we'd just /retest them after the fix lands.  We can
turn the bot back on once we got back to one-external-workaround a
week or so, vs. our current several per day ;).

Docs for the repo: syntax are in [3].

[1]: https://github.com/openshift/installer/pull/415
[2]: https://github.com/openshift/installer/pull/425
[3]: https://help.github.com/articles/searching-issues-and-pull-requests/#search-within-a-users-or-organizations-repositories