Protect all branches of openshift-ansible
All of the jobs for the older branches are set as optional, so it will
be possible to merge PRs into those branches with a LGTM and approval
only, no jobs will need to pass. There is no reason to not protect at
all, when the simple path to merge exists.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>