cluster-launch-installer-e2e: Restore EXIT and TERM signal handlers
I'd removed these in 3f2f01c4
(ci-operator/config/openshift/installer/master: Move to
openshift-install, 2018-09-26, #1677) while removing the shell from
the setup container.  But we got a shell back in b20d571a
(cluster-launch-installer-e2e: Start setting expirationDate again,
2018-09-29, #1761), and restoring these handlers will help:

* Trigger early pod wrap-up when the setup container dies.  Both the
  test and teardown containers are monitoring the 'exit' file so they
  can bail early.

* Forward TERM to the installer, so we can exit gracefully when
  someone TERMs the shell script.

For the signal handling to work, we need to keep the shell process
around, so I've dropped the 'exec' from the openshift-install
invocation and moved it to the background instead (so 'jobs -p' will
list it).