cluster-launch-installer-e2e: Extract image details
Make it easier to debug pod failures by extracting the image metadata
(including the commit ID in io.openshift.build.commit.id).  This makes
it easier to report image bugs upstream and to see whether a
recently-landed change to the source repository was included in your
test run.

With [1], we'd also get the source repository.  But even without
that, having the commit ID and other metadata is useful.

With this commit, the pods.json pull becomes synchronous, because we
need it to complete before we can iterate over pods.json.

[1]: https://github.com/openshift/ci-operator/pull/188