mirror: Create an OCP 4.0 release mirror that will allow builds from RHEL
Repurpose the generic mirror with an OCP 4.0 specific one that uses the
existing mirrors for OCP content under a single repo. After this, any
image build on the cluster can do:

    rm /etc/yum.repos.d/*
    curl http://rpm-ocp-4-0.ci.svc > /etc/yum.repos.d/test.repo
    yum install X

and ensure that they see ONLY the content that an OCP 4.0 build would
see.

This will allow us to experiment with having RHEL dockerfiles checked
in to the source repos that require no change to work against both
CI and OSBS. For instance, the ci-operator could rewrite the Dockerfile
to perform this step in certain circumstances.