cluster-launch-installer-e2e: Block tests on 'create cluster'
Don't run the tests before 'create cluster' exits successfully (if it
fails, the /tmp/shared/exit check will kill the 'test' container).
With this, we no longer have to wait on the shared kubeconfig, because
we know it was created before setup-success.  And once 'create
cluster' successfully blocks on the bootstrap-complete event [1], this
commit will ensure that we run our tests on the production control
plane, and while the bootstrap control plane is still involved.

[1]: https://github.com/openshift/installer/pull/579
     Although we've been seeing some issues with the re-watch logic
     from that pull request.  Once that logic solidifies, we can drop
     the API_UP check as well.