Create a set of base images in CI that can build like OSBS
This creates jobs to maintain base images that have repos that point to
the internal mirrors (so that the images don't have credentials in them).

The base images are available to authenticated users.

The standard-golang.yaml job demonstrates running a build on api.ci that
consumes rhel content. We could make it easy for someone to run a binary
build against the server. The advantage of this approach is that creds
never leave the cluster and CI and user testing is the same.