e2e: Prepare to support new openshift-tests binary in 4.0 jobs
The new binary is simpler and easier to reason about.

Also add a `launch-aws` test for installer that brings a cluster up for
~2 hours to allow debugging failures in a realistic environment. Like
older e2e-gcp-launch jobs, this allows you to access the cluster via
the console to get the secret.

Make test trigger bypass 'all' for jobs that don't run all the time.