ci-operator/templates/openshift: Refactor router-rollout wait (again)
Today I saw [1]:

  error: watch closed before Until timeout
  error openshift-ingress/deploy/router-default did not come up
  sleep: invalid option -- '4'
  Try 'sleep --help' for more information.

I suspect that the 'rollout status' request took long enough that the
fresh 'date' call generated a time larger than wait_expiry_time.  This
commit rerolls the logic last touched by 7991fd30 (Fix how we wait on
router rollout as the new cluster ingress operator, 2018-10-23, #2004).

Now we pick a total wait time (10 minutes), regardless of how many
times we need to reconnect the watcher.  With this commit, each
watcher will try to wait for the full remaining period.  So the first
watcher tries to wait for 10 minutes.  And if the first times out
after 2 minutes, the second watcher will try to wait for 8 minutes.

And the cool-off sleep is no longer parameterized, which removes the
change of flaking like I saw today.

[1]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/688/pull-ci-openshift-installer-master-e2e-aws/1971/build-log.txt