ci-operator/templates/openshift: Drop --log-level=debug
This makes it easier for folks who are less familiar with the debug
logs to find the actual error message.  For folks who *do* need access
to the debug logs, we still have those in the recovered artifacts
since openshift/installer@06c5fd36 (openshift-install: log debug
output to file, 2018-11-15, openshift/installer#689).

Generated with:

  $ sed -i 's/--log-level.debug //' $(git grep -l -- '--log-level.debug')