Allow Origin PR builds in master to inject kubelet into rhcos
The machine-os-content image is injected into our CI stream containing
a snapshot of recent openshift projects (kubelet, crio, oc). To ensure
we are testing what we merge, this PR adds a new layer job to origin
that takes the current machine-os-content image and uses it as input
to a special layering Dockerfile defined in origin which injects the
version of the kubelet and oc binaries from the PR build. The
machine-os-content image is then overriden in stable, but is not
promoted using the new exclusion rule in openshift/ci-operator#254.

This will be the model for crio or any other RPM that doesn't
require special boot time changes (like ignition or the kernel would).
However, this allows us to test upgrades like they will actually
happen in the wild.

This will be merged after the ci-operator change, and after the
dockerfile merges to origin, but before the machine-config-operator
starts respecting the payload image (and causes reboots when
machine start up).