Move all origin builds to publish to ocp/4.0 and switch the base
Our CI strategy is to unify OCP and Origin testing to ensure we have
the best possible coverage of the OCP output. To better control the
availability of prerelease images, settle on the appropriate publishing
location for images as the ocp namespace in the 4.0 stream (which will
become 4.1, 4.2, etc).

Each image in the origin/ocp release is published to a consistent spot
and uses the local RPM mirrors for accessing the appropriate source content.
For most we are switching to use the new UBI RHEL image as our base which
has substantial size improvements. For a few, like
ansible/jenkins/aggregated-logging which have complex dockerfiles, they
have a shim base image that they create for centos that will eventually
be moved to UBI.

The 4.0.0-0.ci release stream will publish the latest ocp/4.0 back to
openshift/origin-v4.0 in the short term to assist in the transition,
although we will eventually drop that target and simply push to quay.