Modify install-config.yaml to use controlPlane and compute instead of machines.
In the cases where the replica count is not customizable, the controlPlane and compute
have been removed to use the defaults for the machine pools.

Update the apiVersion for install-config.yaml to v1beta3, which is the version that
incorporates the controlPlane and compute fields.

See https://github.com/openshift/installer/pull/1157