cluster/ci/config/prow/plugins: Enable the override plugin
The plugin is defined in [1].  This allows repo admins to force
specific tests green as a slightly less invasive alternative to
directly merging PRs.

I've gone through and enabled it for every block that enabled the very
important shrug plugin ;), since that seemed like a good way to find
entries where the plugin list was starting from scratch.

[1]: https://github.com/kubernetes/test-infra/tree/master/prow/plugins/override