Do not block merges on unknown contexts in openshift/release
We do not have any non-Prow jobs set up in openshift/release, but
now we have rehearsals, which are unfortunately considered as unknown
contexts by Prow (because while they are still Prow jobs, they are not
*configured* for the repository, and a lot of things, including Tide,
are driven by the sets of jobs configured for a repo). If we want to
have rehearsal non-blocking, we need to not block on unknown contexts.