ci-operator/templates/openshift: Use IMAGE_TESTS for teardown gzip
The teardown container requires oc (for retrieving cluster logs), gzip
(for compressing those logs), and openshift-install (for destroying
the cluster).  We were already copying in oc from another container;
with this commit we copy in the statically-linked openshift-install
too.  That leaves gzip, which we will be built in to IMAGE_TESTS as "a
command needed in the CI-test context".  Using IMAGE_TESTS for
teardown means we no longer need to copy oc into the shared volume for
the teardown container (although we may need to copy it in for other
containers, e.g. the src template's test container).

I've also dropped an unnecessary /bin/ prefix from the 'create
cluster' invocations, which we'd been dragging around from the
pre-shell days of 3f2f01c4
(ci-operator/config/openshift/installer/master: Move to
openshift-install, 2018-09-26, #1677).