Add a presubmit job to test upgrades on the CVO repo
Requires openshift/ci-operator#284 before it will function. Will be
used as the template for other jobs.

The job launches the cluster using a release image build from the
current state of the integration stream (release:initial). It then
launches the upgrade tests with

    openshift-tests run-upgrade all --to-image=.../ci-op-.../release:latest

The test job will start the prereq tests, then once they are stable, launch
the upgrade, monitoring invariants while the test continues. Once the upgrade
is complete or fails, the postcondition tests will run and then the cluster
is torn down.

Due to limitations in the upstream infrastructure for upgrade tests we don't
get the nice separation between tests that e2e runs do which is a future
improvement.

After this test is working we'll add it to other repos as a standard job,
including the release verification jobs. We'll also invoke it from the
release controller dynamically to calculate upgrade edges prior to
promoting a job.