openstack: minor fixes for openstack-e2e
We were getting OOMKilled pods with the m1.medium flavor, m1.xlarge
has 16GB of RAM and does not have as much issue with OOMKilled pods.

We need to use the image with `openshift-tests` binary for running
the test container.

We also want to use a smaller subset of tests while we get the job
stable. (conformance/parallel/minimal)