machine-config-operator: Make all tests run when requested
Today we spawn *three* clusters every time someone does a `git push`.
That's pretty nuts, and IMO doing this type of thing across
all the repos is both an unnecessary waste of money, and also
exacerbates issues with AWS resource limits.

I sometimes hesitate at doing `git push` to fix a typo in a comment
as part of a PR review because I know that push may end up
stealing an AWS NAT limit that's going to be used by an actually
important job.

For the MCO specifically many of our changes are extremely
unlikely to break `e2e-aws`, and if they did they'd break
`e2e-aws-op` too.

So we'll do e.g. `/test e2e-aws` on demand in PRs, or
`/test all`, etc.