openstack: Fix concurrency bugs
We needed to switch from the aws default for CLUSTER_NAME, because
it was too long and was getting truncated. Unfortunately, we picked
the non-unique half. Any jobs with the same name will all try to hit
the same API URL.

Instead, we can use the CI namespace, since this will be unique even
for two concurrent runs of the same job.

This patch also fixes an issue where jobs run at nearly the same time
could steal the FIP. This is because we check if a FIP is not used, but
then do not actually attach to a VM for a few minutes. A second job that
checks that FIP during that window will also see it as not in use.

Instead, we just always get a new FIP and make sure to delete it in cleanup
so we don't exhaust all the FIPs.