openstack: run openstack-e2e job on every installer patch
This changes the openstack-e2e job to run on every patch to the
installer. The job will not block merging.

Running on every patch will help us to find any bugs in our CI
infrastructure related to running so many jobs concurrently. It
will also give us some data about the actual pass rate of the job.
This will be useful to iterate and improve the job.