ci-operator/jobs/openshift/installer/openshift-installer-master-presubmits: Whitelist passing libvirt tests
As of now not all parallel tests pass on libvirt like they do on AWS.
With this commit we whitelist known-good tests so we can start gating
on libvirt success.  As libvirt support improves, we can expand the
list until we have full coverage.

Use EOTESTS instead of the conventional EOF because this is getting
nested inside the EOF-delimited here-document in the
cluster-launch-installer-libvirt-e2e.yaml template.