Create a periodic that upgrades from the latest stable to 4.2
This periodic will be a template to verify upgrades from 4.1 to 4.2.
It uses the release controller to identify which versions are available
and then tests the upgrade between 4-stable in 4.1.z to the latest
accepted 4.2.0-0.ci. Future jobs will use this as a template.

It sets the annotations the release controller uses to verify edges
so that the graph will report upgrades from 4.1.0