Revert "cluster/ci/config/prow/openshift/artifact-uploader: Map .log -> text/plain"
This reverts commit c850713c34a45f5000c211eb2663522d8abef198.

Apparently I got the wrong place, and this one is for some
experimental feature running an old container image (which does not
support -extensions yet) for uploading content to GCS from containers
as they finish (instead of waiting for the whole Pod?).  Anyhow,
rolling this back until the backing container image gets bumped to
pick up the new option.