Unprotect branches used by nightly CI
These repos all use a script to synch the release-next branch to
upstream/master + the files needed to enable ci-operator and force push
the changes. A simple change is made on the release-next-ci branch to
trigger the CI, and a force push is used there in case a PR already
exists. This change disables branch protection on these two branches for
the Knative repos that use this automation. No releases are made from
these branches, so this shouldn't violate any RCM rules. This is just to
be prepared for any breakage caused by upstream changes.