Force OpenShiftSDN on openstack-e2e runs
Currently work on support for Kuryr-Kubernetes is ongoing on
cluster-network-operator, as well as openshift/installer. The point is
openshift/installer is supposed to automatically configure Kuryr when
Octavia and trunk ports are enabled on OpenStack installation.
openstack-e2e runs on such cloud and merging a commit to
openshift/installer that enables Kuryr would switch this job to use it
as well, possibly making it unstable.

In order to avoid that this commit is forcing openstack-e2e to generate
manifests first (using `openshift-install create manifests`), then
modify cluster-network-config to configure OpenShiftSDN and only then
run the rest of the installation. Kuryr may be allowed only when
KURYR_ENABLED is set to "true".